Menu item is a separate component now
